requirements.txt
# ── Web Framework ──────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Frontend ───────────────────────────────────────────────────────────────
streamlit==1.35.0

# ── Computer Vision ────────────────────────────────────────────────────────
opencv-python==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
imagehash==4.3.1
scikit-image==0.23.2

# ── Google AI ──────────────────────────────────────────────────────────────
google-generativeai==0.7.2

# ── Firebase ───────────────────────────────────────────────────────────────
firebase-admin==6.5.0

# ── Utilities ──────────────────────────────────────────────────────────────
python-dotenv==1.0.1
requests==2.31.0
aiofiles==23.2.1
pydantic==2.7.1
